+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.1 [2025-10-13]
 
 - Updated Globe component to isolate internal styling
+- Removed Cesium Ion dependencies within Components
 
 ## 1.1.0 [2025-10-13]